🔧 Configure Storybook to use Chakra UI
