+++ .storybook/main.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const toPath = (_path) => path.join(process.cwd(), _path);
 const TsconfigPathsPlugin = require('tsconfig-paths-webpack-plugin');
 
 module.exports = {
@@ -12,6 +13,11 @@ module.exports = {
     config.resolve.modules = [...(config.resolve.modules || []), path.resolve(__dirname, '../src')];
     config.resolve.plugins = [...(config.resolve.plugins || []), new TsconfigPathsPlugin()];
     config.resolve.fallback = { fs: false, path: false };
+    config.resolve.alias = {
+      ...(config.resolve.alias || []),
+      '@emotion/core': toPath('node_modules/@emotion/react'),
+      'emotion-theming': toPath('node_modules/@emotion/react'),
+    };
 
     return config;
   },
